fixs

+++ src/screens/Settings/index.tsx
@@ -1,23 +1,84 @@
-import { Box, Button, Divider, Heading, Text, VStack } from "native-base";
+import { Box, Button, Text, VStack, useToast } from "native-base";
 import React from "react";
 import { SafeAreaView } from "react-native-safe-area-context";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { useGetProducts, useGetDaysToExpire } from "../../hooks";
+import { Alert } from "react-native";
 
 export const Settings: React.FC = () => {
   const getProducts = useGetProducts();
+  const [loading, setLoading] = React.useState<{
+    expired: boolean;
+    all: boolean;
+  }>({ expired: false, all: false });
+  const alert = useToast();
 
-  const handleDeleteExpired = async () => {
-    let producsData = await getProducts();
-    let producsFilter = producsData.filter(
-      (productData) =>
-        useGetDaysToExpire(new Date(productData.expirationDate)) > 0
+  const handleDeleteExpired = () => {
+    Alert.alert(
+      "Delete Expired Products",
+      "Are you sure you want to delete all expired products?",
+      [
+        {
+          text: "Cancel",
+          onPress: () => {},
+          style: "cancel",
+        },
+        {
+          text: "Delete",
+          onPress: async () => {
+            setLoading({ ...loading, expired: true });
+            let producsData = await getProducts();
+            let producsFilter = producsData.filter(
+              (productData) =>
+                useGetDaysToExpire(new Date(productData.expirationDate)) > 0
+            );
+            await AsyncStorage.setItem(
+              "products",
+              JSON.stringify(producsFilter)
+            );
+            alert.show({
+              pl: 2,
+              pr: 2,
+              title: "All expired products deleted",
+              placement: "top",
+              bgColor: "black",
+              borderRadius: 13,
+            });
+            setLoading({ ...loading, expired: false });
+          },
+        },
+      ]
     );
-    await AsyncStorage.setItem("products", JSON.stringify(producsFilter));
   };
 
-  const handleDeleteAll = async () => {
-    await AsyncStorage.setItem("products", JSON.stringify([]));
+  const handleDeleteAll = () => {
+    Alert.alert(
+      "Dellete All Products",
+      "Are you sure you want to delete all products?",
+      [
+        {
+          text: "Cancel",
+          onPress: () => {},
+          style: "cancel",
+        },
+        {
+          text: "Delete",
+          onPress: async () => {
+            setLoading({ ...loading, all: true });
+            await AsyncStorage.setItem("products", JSON.stringify([]));
+            alert.show({
+              pl: 2,
+              pr: 2,
+              title: "All products deleted",
+              placement: "top",
+              bgColor: "black",
+              borderRadius: 13,
+            });
+            setLoading({ ...loading, all: false });
+          },
+        },
+      ]
+    );
   };
 
   return (
@@ -35,6 +96,7 @@ export const Settings: React.FC = () => {
       </Box>
       <VStack alignItems={"center"} space={5}>
         <Button
+          isLoading={loading.expired}
           onPress={handleDeleteExpired}
           w="60%"
           bg="black"
@@ -45,6 +107,7 @@ export const Settings: React.FC = () => {
 
         <Button
           onPress={handleDeleteAll}
+          isLoading={loading.all}
           w="60%"
           bg="black"
           colorScheme={"light"}
@@ -55,22 +118,3 @@ export const Settings: React.FC = () => {
     </SafeAreaView>
   );
 };
-
-/* {
-  "cli": {
-    "version": ">= 0.49.0"
-  },
-  "build": {
-    "development": {
-      "developmentClient": true,
-      "distribution": "internal"
-    },
-    "preview": {
-      "distribution": "internal"
-    },
-    "production": {}
-  },
-  "submit": {
-    "production": {}
-  }
-} */
